backend/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-cors-headers==2.5.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.9.2
geographiclib==1.49
geopy==1.19.0
googlemaps==2.4
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
oauthlib==3.0.1
openapi-codec==1.3.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
